==> notebooks/02_gdelt_cleaning.ipynb <==
import pandas as pd

gdelt = pd.read_parquet('../cache/gdelt.parquet')

gdelt.head()

gdelt.columns

# convert categorical columns to category type
# gdelt.Actor1Country = gdelt.Actor1Country.astype('category')
# gdelt.Actor1GeoCountry = gdelt.Actor1GeoCountry.astype('category')
# gdelt.Actor1Type = gdelt.Actor1Type.astype('category')
# gdelt.Actor2Country = gdelt.Actor2Country.astype('category')
# gdelt.Actor2GeoCountry = gdelt.Actor2GeoCountry.astype('category')
# gdelt.Actor2Type = gdelt.Actor2Type.astype('category')
# gdelt.ActionCountry = gdelt.ActionCountry.astype('category')
# gdelt.EventRootCode = gdelt.EventRootCode.astype('category')
# gdelt.QuadClass = gdelt.QuadClass.astype('category')

# datetime to unix timestamp
gdelt['Timestamp'] = gdelt.Date.astype('int64') // 10**9

gdelt.dtypes

# replace quadclass nums with meaningful names
gdelt['QuadClass'] = gdelt['QuadClass'].map({
    1: 'VerbalCoop', 
    2: 'MaterialCoop', 
    3: 'VerbalConf', 
    4: 'MaterialConf'
})

# replace event root code nums with meaningful names and rename to EventType

# first convert to numeric
gdelt['EventRootCode'] = pd.to_numeric(gdelt['EventRootCode'], errors='coerce')
gdelt['EventRootCode'] = gdelt['EventRootCode'].map({
    1: 'Statement',
    2: 'Appeal',
    3: 'Intent to Coop',
    4: 'Consult',
    5: 'Diplom Coop',
    6: 'Material Coop',
    7: 'Aid',
    8: 'Yield',
    9: 'Investigate',
    10: 'Demand',
    11: 'Disapprove',
    12: 'Reject',
    13: 'Threaten',
    14: 'Protest',
    15: 'Demonstrate Force',
    16: 'Reduce Relations',
    17: 'Coerce',
    18: 'Assault',
    19: 'Fight',
    20: 'Mass Violence'
})
gdelt.rename(columns={'EventRootCode': 'EventType'}, inplace=True)

# convert back to category
# gdelt.EventType = gdelt.EventType.astype('category')

# there are 3 NaN values in EventType, so we can drop them
gdelt['EventType'].value_counts(dropna=False)[::-1]

gdelt.dropna(subset=['EventType'], inplace=True)

# generalize actor types
actor_types = {
    'GOV': ['GOV', 'COP', 'MIL', 'JUD', 'SPY', 'ELI', 'LEG'],
    'Opposition': ['OPP', 'INS', 'REB', 'SEP', 'CRM'],
    'Business': ['BUS', 'AGR', 'DEV', 'LAB'],
    'Benevolent': ['EDU', 'MED', 'HRI', 'HLH', 'ENV'],
}

def generalize_actor_type(actor_type):
    for k, v in actor_types.items():
        if actor_type in v:
            return k
    return 'Other'

gdelt['Actor1TypeGeneral'] = gdelt['Actor1Type'].apply(generalize_actor_type)
gdelt['Actor2TypeGeneral'] = gdelt['Actor2Type'].apply(generalize_actor_type)

gdelt['Actor1TypeGeneral'].value_counts()

gdelt['Actor2TypeGeneral'].value_counts()

# we can see that GoldsteinScale has 23 NaN values
# this is a very small number, so we can drop them
gdelt[gdelt['GoldsteinScale'].isna()].__len__()

gdelt.dropna(subset=['GoldsteinScale'], inplace=True)

# # replace countries not in the top 20 with 'Other'
# top_countries = gdelt['Actor1Country'].value_counts().head(20).index
# gdelt['Actor1Country'] = gdelt['Actor1Country'].apply(lambda x: x if x in top_countries else 'Other')
# gdelt['Actor2Country'] = gdelt['Actor2Country'].apply(lambda x: x if x in top_countries else 'Other')

# # replace geo countries not in the top 20 with 'Other'
# top_geo_countries = gdelt['Actor1GeoCountry'].value_counts().head(20).index
# gdelt['Actor1GeoCountry'] = gdelt['Actor1GeoCountry'].apply(lambda x: x if x in top_geo_countries else 'Other')
# gdelt['Actor2GeoCountry'] = gdelt['Actor2GeoCountry'].apply(lambda x: x if x in top_geo_countries else 'Other')

# # replace action countries not in the top 20 with 'Other'
# top_action_countries = gdelt['ActionCountry'].value_counts().head(20).index
# gdelt['ActionCountry'] = gdelt['ActionCountry'].apply(lambda x: x if x in top_action_countries else 'Other')

gdelt[gdelt.NumArticles > 500].shape[0]

There are only 99 events with more than 500 articles, so we can cap the number of articles at 500 to avoid outliers.

# cap number of articles at 500
gdelt['NumArticles'] = gdelt['NumArticles'].clip(upper=500)

# reorder columns
gdelt = gdelt[['Date', 'Timestamp',  # Date
    'Actor1Country', 'Actor1GeoCountry', 'Actor1Type', 'Actor1TypeGeneral',  # Actor1
    'Actor2Country', 'Actor2GeoCountry', 'Actor2Type', 'Actor2TypeGeneral',  # Actor2
    'ActionCountry', 'EventType', 'QuadClass', 'GoldsteinScale',  # Event
    'NumSources', 'NumArticles', 'AvgTone', 'Source']]  # Reaction

from funcs.advanced_describe import advanced_describe

advanced_describe(gdelt, drop_cols=['Date', 'Timestamp', 'Source'])

gdelt.head()

After analyzing impact of different features on btc price, we can drop some columns and remove some less frequent values.

gdelt.columns

gdelt = gdelt.drop(columns=['Source', 'Actor1TypeGeneral', 'Actor2TypeGeneral', 'QuadClass'])

gdelt.head()

# save cleaned gdelt
gdelt.to_parquet('../cache/gdelt_cleaned.parquet')
