==> notebooks/04_merging_datasets.ipynb <==
import pandas as pd
gdelt = pd.read_parquet('../cache/gdelt_cleaned.parquet')
btc = pd.read_parquet('../cache/btc_cleaned.parquet')

include_impact = True
if include_impact:
    from funcs.impact_scoring import calc_event_magnitude, calc_event_impact, get_impact_bin
    gdelt['Magnitude'] = gdelt.apply(calc_event_magnitude, axis=1)
    print('Magnitude calculated')
    
    gdelt['Impact'] = gdelt.apply(calc_event_impact, axis=1)
    print('Impact calculated')
    
    gdelt['Impact_bin'] = gdelt['Impact'].apply(get_impact_bin)
    print('Impact bin calculated')

# merge the two dataframes
gdelt = gdelt.sort_values('Timestamp')
btc = btc.sort_values('Timestamp')

merged = pd.merge_asof(gdelt, btc, on='Timestamp')
merged.head()

merged.dropna().sample().iloc[0]

from funcs.advanced_describe import advanced_describe
advanced_describe(merged)

merged.to_parquet("../cache/merged.parquet")
