==> notebooks/07_encoding_data.ipynb <==
import pandas as pd

merged_0q = pd.read_parquet('../cache/merged_0q.parquet')
merged_50q = pd.read_parquet('../cache/merged_50q.parquet')
merged_90q = pd.read_parquet('../cache/merged_90q.parquet')
merged_99q = pd.read_parquet('../cache/merged_99q.parquet')

merged_0q.dtypes

merged_0q.columns

numerical_columns = merged_0q.select_dtypes(include=['int64', 'float64']).columns.tolist()
for col in numerical_columns:
    print(col, merged_0q[col].nunique())

categorical_columns = merged_0q.select_dtypes(include=['object', 'category']).columns.tolist()
for col in categorical_columns:
    print(col, merged_0q[col].nunique())

from sklearn.preprocessing import LabelEncoder, StandardScaler
def prep_data(df,
              catg_cols=('Actor1Country', 'Actor1GeoCountry', 'Actor1Type', 'Actor2Country', 'Actor2GeoCountry', 'Actor2Type', 'ActionCountry', 'EventType'),
              num_cols=('AvgTone', 'GoldsteinScale', 'NumSources', 'NumArticles')):
    encoders = {}
    for catg_col in catg_cols:
        encoders[catg_col] = LabelEncoder()
        df[f'{catg_col}_enc'] = encoders[catg_col].fit_transform(df[catg_col])
    
    num_cols = list(num_cols) + [f'{feature}_enc' for feature in catg_cols]
    scaler = StandardScaler()
    df[num_cols] = scaler.fit_transform(df[num_cols])
    
    return df, encoders, scaler, num_cols

df_0q_encoded, encoders_0q, scaler_0q, num_cols_0q = prep_data(merged_0q)
df_50q_encoded, encoders_50q, scaler_50q, num_cols_50q = prep_data(merged_50q)
df_90q_encoded, encoders_90q, scaler_90q, num_cols_90q = prep_data(merged_90q)
df_99q_encoded, encoders_99q, scaler_99q, num_cols_99q = prep_data(merged_99q)

df_0q_encoded.head()

dfs = [df_0q_encoded, df_50q_encoded, df_90q_encoded, df_99q_encoded]
num_cols = [num_cols_0q, num_cols_50q, num_cols_90q, num_cols_99q]
for i, df in enumerate(dfs):
    df.set_index('Date', inplace=True)
    # keep only numerical + btc price columns
    df = df[num_cols[i] + ['pct_change_30min']]
    dfs[i] = df
df_0q, df_50q, df_90q, df_99q = dfs

df_0q.head()

# convert bool columns to int (currently not needed)
# bool_columns = merged_encoded.select_dtypes(include='bool').columns
# merged_encoded[bool_columns] = merged_encoded[bool_columns].astype(int)

df_0q.to_parquet('../cache/prepd_0q.parquet')
df_50q.to_parquet('../cache/prepd_50q.parquet')
df_90q.to_parquet('../cache/prepd_90q.parquet')
df_99q.to_parquet('../cache/prepd_99q.parquet')
